Make SearchIndex.Search accepts query as string instead of *Document
